+++ public/types/src/scripts/choices.d.ts
@@ -11,7 +11,7 @@ import { PassedElementType } from './interfaces';
 import { Choices as ChoicesInterface } from './interfaces/choices';
 import { EventChoice } from './interfaces/event-choice';
 import { Templates } from './interfaces/templates';
-import { SearchHandler } from './interfaces/search';
+import { Searcher } from './interfaces/search';
 /**
  * Choices
  * @author Josh Johnson<josh@joshuajohnson.co.uk>
@@ -62,7 +62,7 @@ declare class Choices implements ChoicesInterface {
     _presetChoices: (ChoiceFull | GroupFull)[];
     _presetItems: ChoiceFull[];
     _initialItems: string[];
-    _searchFn: SearchHandler;
+    _searcher: Searcher<ChoiceFull>;
     constructor(element?: string | Element | HTMLInputElement | HTMLSelectElement, userConfig?: Partial<Options>);
     init(): void;
     destroy(): void;