+++ public/types/src/scripts/search/index.d.ts
@@ -1,3 +1,3 @@
-import { SearchHandler } from '../interfaces/search';
-declare let search: SearchHandler;
-export default search;
+import { Options } from '../interfaces';
+import { Searcher } from '../interfaces/search';
+export declare function getSearcher<T extends object>(config: Options): Searcher<T>;