+++ src/scripts/search/prefix-filter.ts
@@ -1,31 +1,48 @@
 import { Options } from '../interfaces';
-import { SearchResult } from '../interfaces/search';
+import { Searcher, SearchResult } from '../interfaces/search';
 
-export function searchByPrefixFilter<T extends object>(
-  config: Options,
-  haystack: T[],
-  _needle: string,
-): SearchResult<T>[] {
-  const fields = config.searchFields;
-  if (!fields || fields.length === 0 || _needle === '') {
-    return [];
+export class SearchByPrefixFilter<T extends object> implements Searcher<T> {
+  _fields: string[];
+
+  _haystack: T[] = [];
+
+  constructor(config: Options) {
+    this._fields = config.searchFields;
   }
 
-  const needle = _needle.toLowerCase();
+  index(data: T[]): void {
+    this._haystack = data;
+  }
 
-  return haystack
-    .filter((obj) =>
-      fields.some(
-        (field) =>
-          field in obj &&
-          (obj[field] as string).toLowerCase().startsWith(needle),
-      ),
-    )
-    .map((value, index): SearchResult<T> => {
-      return {
-        item: value,
-        score: index,
-        rank: index,
-      };
-    });
+  reset(): void {
+    this._haystack = [];
+  }
+
+  isEmptyIndex(): boolean {
+    return this._haystack.length === 0;
+  }
+
+  search(_needle: string) {
+    const fields = this._fields;
+    if (!fields || fields.length === 0 || _needle === '') {
+      return [];
+    }
+    const needle = _needle.toLowerCase();
+
+    return this._haystack
+      .filter((obj) =>
+        fields.some(
+          (field) =>
+            field in obj &&
+            (obj[field] as string).toLowerCase().startsWith(needle),
+        ),
+      )
+      .map((value, index): SearchResult<T> => {
+        return {
+          item: value,
+          score: index,
+          rank: index,
+        };
+      });
+  }
 }