+++ test/scripts/search/index.test.ts
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
 import { beforeEach } from 'vitest';
-import { searchByPrefixFilter } from '../../../src/scripts/search/prefix-filter';
-import { searchByFuse } from '../../../src/scripts/search/fuse';
 import { DEFAULT_CONFIG } from '../../../src';
 import { cloneObject } from '../../../src/scripts/lib/utils';
+import { SearchByFuse } from '../../../src/scripts/search/fuse';
+import { SearchByPrefixFilter } from '../../../src/scripts/search/prefix-filter';
 
 export interface SearchableShape {
   label: string;
@@ -21,29 +21,28 @@ describe('search', () => {
   );
 
   describe('fuse-full', () => {
-    const search = searchByFuse;
+    let searcher: SearchByFuse<SearchableShape>;
+
     beforeEach(() => {
       process.env.SEARCH_FUSE = 'full';
+      searcher = new SearchByFuse<SearchableShape>(options);
+      searcher.index(haystack);
     });
     it('empty result', () => {
-      const results = search<SearchableShape>(options, haystack, '');
+      const results = searcher.search('');
       expect(results.length).eq(0);
     });
     it('label prefix', () => {
-      const results = search<SearchableShape>(options, haystack, 'label');
+      const results = searcher.search('label');
       expect(results.length).eq(haystack.length);
     });
     it('label suffix', () => {
-      const results = search<SearchableShape>(
-        options,
-        haystack,
-        `${haystack.length - 1}`,
-      );
+      const results = searcher.search(`${haystack.length - 1}`);
       expect(results.length).eq(1);
     });
 
     it('search order', () => {
-      const results = search<SearchableShape>(options, haystack, 'label');
+      const results = searcher.search('label');
 
       expect(results.length).eq(haystack.length);
       haystack.forEach((value, index) => {
@@ -60,7 +59,10 @@ describe('search', () => {
 
         return a.score < b.score ? 1 : -1;
       };
-      const results = search<SearchableShape>(opts, haystack, 'label');
+
+      searcher = new SearchByFuse<SearchableShape>(opts);
+      searcher.index(haystack);
+      const results = searcher.search('label');
 
       expect(results.length).eq(haystack.length);
       haystack.reverse().forEach((value, index) => {
@@ -70,28 +72,26 @@ describe('search', () => {
   });
 
   describe('fuse-basic', () => {
-    const search = searchByFuse;
+    let searcher: SearchByFuse<SearchableShape>;
     beforeEach(() => {
       process.env.SEARCH_FUSE = 'basic';
+      searcher = new SearchByFuse<SearchableShape>(options);
+      searcher.index(haystack);
     });
     it('empty result', () => {
-      const results = search<SearchableShape>(options, haystack, '');
+      const results = searcher.search('');
       expect(results.length).eq(0);
     });
     it('label prefix', () => {
-      const results = search<SearchableShape>(options, haystack, 'label');
+      const results = searcher.search('label');
       expect(results.length).eq(haystack.length);
     });
     it('label suffix', () => {
-      const results = search<SearchableShape>(
-        options,
-        haystack,
-        `${haystack.length - 1}`,
-      );
+      const results = searcher.search(`${haystack.length - 1}`);
       expect(results.length).eq(1);
     });
     it('search order', () => {
-      const results = search<SearchableShape>(options, haystack, 'label');
+      const results = searcher.search('label');
 
       expect(results.length).eq(haystack.length);
       haystack.forEach((value, index) => {
@@ -101,22 +101,22 @@ describe('search', () => {
   });
 
   describe('prefix-filter', () => {
-    const search = searchByPrefixFilter;
-    process.env.SEARCH_FUSE = undefined;
+    let searcher: SearchByPrefixFilter<SearchableShape>;
+    beforeEach(() => {
+      process.env.SEARCH_FUSE = undefined;
+      searcher = new SearchByPrefixFilter<SearchableShape>(options);
+      searcher.index(haystack);
+    });
     it('empty result', () => {
-      const results = search<SearchableShape>(options, haystack, '');
+      const results = searcher.search('');
       expect(results.length).eq(0);
     });
     it('label prefix', () => {
-      const results = search<SearchableShape>(options, haystack, 'label');
+      const results = searcher.search('label');
       expect(results.length).eq(haystack.length);
     });
     it('label suffix', () => {
-      const results = search<SearchableShape>(
-        options,
-        haystack,
-        `${haystack.length - 1}`,
-      );
+      const results = searcher.search(`${haystack.length - 1}`);
       expect(results.length).eq(0);
     });
   });